requirements.txt
bcrypt==4.0.1
bleach==6.0.0
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Login==0.6.2
flask-mailman==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mkdocs-material-extensions==1.1.1
passlib==1.7.4
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.0.1
